requirements.txt
streamlit
python-docx
PyPDF2
language-tool-python
textsta